requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.7.0
requests>=2.31.0
tqdm>=4.66.0
